tests/wpt/tests/html/canvas/offscreen/canvas-host/2d.canvas.host.type.name.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.host.type.name</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.type.name</h1>
<p class="desc">OffscreenCanvas type and toString</p>


<script>
var t = async_test("OffscreenCanvas type and toString");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assertSame(Object.prototype.toString.call(canvas), '[object OffscreenCanvas]', "Object.prototype.toString.call(canvas)", "'[object OffscreenCanvas]'");
  t.done();

});
</script>
